set up of transformation module

+++ src/main/java/ch/ethz/idsc/demo/mg/pipeline/PhysicalBlob.java
@@ -4,6 +4,6 @@ package ch.ethz.idsc.demo.mg.pipeline;
 // provides blob object in physical space
 public class PhysicalBlob {
   // fields
-  private float[] pos; // [m] in gokart reference frame
-  private float[] vel;
+  private double[] pos; // [m] in gokart reference frame
+  private double[] vel;
 }
